@@ -21,7 +21,7 @@ class PostRemoteDataSourceImpl implements PostRemoteDataSource {
   @override
   Future<List<PostModel>> getAllPosts() async {
     final response = await client.get(
-      Uri.parse(BASE_URL + "/posts/"),
+      Uri.parse("$BASE_URL/posts/"),
       headers: {"Content-Type": "application/json"},
     );
 
@@ -45,7 +45,7 @@ class PostRemoteDataSourceImpl implements PostRemoteDataSource {
     };
 
     final response =
-        await client.post(Uri.parse(BASE_URL + "/posts/"), body: body);
+        await client.post(Uri.parse("$BASE_URL/posts/"), body: body);
 
     if (response.statusCode == 201) {
       return Future.value(unit);
@@ -57,7 +57,7 @@ class PostRemoteDataSourceImpl implements PostRemoteDataSource {
   @override
   Future<Unit> deletePost(int postId) async {
     final response = await client.delete(
-      Uri.parse(BASE_URL + "/posts/${postId.toString()}"),
+      Uri.parse("$BASE_URL/posts/${postId.toString()}"),
       headers: {"Content-Type": "application/json"},
     );
 
@@ -77,7 +77,7 @@ class PostRemoteDataSourceImpl implements PostRemoteDataSource {
     };
 
     final response = await client.patch(
-      Uri.parse(BASE_URL + "/posts/$postId"),
+      Uri.parse("$BASE_URL/posts/$postId"),
       body: body,
     );
 
