@@ -14,7 +14,7 @@ class LoadingPostsState extends PostsState {}
 class LoadedPostsState extends PostsState {
   final List<Post> posts;
 
-  LoadedPostsState({required this.posts});
+  const LoadedPostsState({required this.posts});
 
   @override
   List<Object> get props => [posts];
@@ -23,7 +23,7 @@ class LoadedPostsState extends PostsState {
 class ErrorPostsState extends PostsState {
   final String message;
 
-  ErrorPostsState({required this.message});
+  const ErrorPostsState({required this.message});
 
   @override
   List<Object> get props => [message];
