@@ -14,7 +14,7 @@ class TextFormFieldWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-        padding: EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+        padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
         child: TextFormField(
           minLines: multiLines ? 6 : 1,
           maxLines: multiLines ? 6 : 1,
