@@ -9,13 +9,13 @@ class MessageDisplayWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
+    return SizedBox(
       height: MediaQuery.of(context).size.height / 3,
       child: Center(
           child: SingleChildScrollView(
         child: Text(
           message,
-          style: TextStyle(fontSize: 25),
+          style: const TextStyle(fontSize: 25),
           textAlign: TextAlign.center,
         ),
       )),
